Add files via upload
These are the main file with the calculator class and a file with 19 tests. I have made function power with the feature: if it gets more than two numbers from the input, it uses only first two of them to prevent counting enormously large numbers by a mistake. For example, power(2, 3, 999) = power(2, 3).
Also the programm doesn't always complete testing because of the system rounding. I think it is difficult to do anything with this, especially it is easy to look, where did the programm fail and to make sure that number are really similar to each other. Anyway, accuracy is really high.